requirements.txt
requests
textblob
tweepy
langdetect
nltk
sklearn
python-dotenv